feature/era-glonass-integration-r2:
moved all type-safety helpers based on `Generic`s from
__carma-era-glonass-integration__ package to __carma-utils__
