+++ carma-era-glonass-integration/src/Carma/EraGlonass/Types/EGUpdateCallCardAcceptCode.hs
@@ -15,7 +15,7 @@ import           Data.Swagger
 
 import           Carma.Utils.StringyEnum
 import           Carma.Utils.StringyEnum.SwaggerSchema
-import           Carma.EraGlonass.Types.Helpers
+import           Carma.Utils.TypeSafe.Generic.DataType
 
 
 data EGUpdateCallCardAcceptCode
